+++ PlaygroundLite/PlaygroundLite/ViewModels/LinearViewModel.cs
@@ -4,12 +4,11 @@ namespace PlaygroundLite.ViewModels
 {
     public class LinearViewModel : GradientViewModel<LinearGradient>
     {
-        private float _angle;
-        public float Angle
+        private double _angle;
+        public double Angle
         {
             get => _angle;
-            set => SetProperty(ref _angle, value, 
-                onChanged: () => Gradient.Angle = _angle);
+            set => SetProperty(ref _angle, value, onChanged: () => Gradient.Angle = _angle);
         }
 
         public LinearViewModel()
